test: includes calculation tests
